V1.1.1

See merge request !2

+++ src/core/lib/utils/hashutil.h
@@ -30,6 +30,7 @@
 #include <iostream>
 #include <string>
 #include <vector>
+#include <stdexcept>
 using std::string;
 using std::vector;
 
@@ -39,7 +40,7 @@ enum HashAlgorithm { SHA_256 = 0, SHA_512 = 1 };
 
 class HashUtil {
 public:
-	static void Hash(string message, HashAlgorithm algo, vector<uint8_t>& digest) {
+	static void Hash(string message, HashAlgorithm algo, vector<int64_t>& digest) {
 		switch (algo) {
 		case SHA_256:
 			SHA256(message, digest);
@@ -58,8 +59,8 @@ public:
 	static std::string HashString(std::string message);
 
 private:
-	static void SHA256(string message, vector<uint8_t>& digest);
-	static void SHA512(string message, vector<uint8_t>& digest);
+	static void SHA256(string message, vector<int64_t>& digest);
+	static void SHA512(string message, vector<int64_t>& digest);
 	static const uint32_t k_256[64];
 	static const uint64_t k_512[80];
 };
